opendata_module/test_requirements.txt
coverage
pytest
pytest-mock
pytest-ruff
pytest-html
ruff
mongomock
psycopg2-binary
pytest-django
python-dateutil
freezegun
django-stubs
pytz
pyyaml